+++ nodejs/eks/workerpool.ts
@@ -16,51 +16,9 @@ import * as aws from "@pulumi/aws";
 import * as k8s from "@pulumi/kubernetes";
 import * as pulumi from "@pulumi/pulumi";
 import * as crypto from "crypto";
-import * as jsyaml from "js-yaml";
 
 import { VpcCni } from "./cni";
-import { ServiceRole } from "./servicerole";
 import transform from "./transform";
-
-/**
- * RoleMapping describes a mapping from an AWS IAM role to a Kubernetes user and groups.
- */
-export interface RoleMapping {
-    /**
-     * The ARN of the IAM role to add.
-     */
-    roleArn: pulumi.Input<aws.ARN>;
-
-     /**
-     * The user name within Kubernetes to map to the IAM role. By default, the user name is the ARN of the IAM role.
-     */
-    username: pulumi.Input<string>;
-
-     /**
-     * A list of groups within Kubernetes to which the role is mapped.
-     */
-    groups: pulumi.Input<pulumi.Input<string>[]>;
-}
-
- /**
- * UserMapping describes a mapping from an AWS IAM user to a Kubernetes user and groups.
- */
-export interface UserMapping {
-    /**
-     * The ARN of the IAM user to add.
-     */
-    userArn: pulumi.Input<aws.ARN>;
-
-     /**
-     * The user name within Kubernetes to map to the IAM user. By default, the user name is the ARN of the IAM user.
-     */
-    username: pulumi.Input<string>;
-
-     /**
-     * A list of groups within Kubernetes to which the user is mapped to.
-     */
-    groups: pulumi.Input<pulumi.Input<string>[]>;
-}
 
 /**
  * WorkerPoolOptions describes the configuration options accepted by a WorkerPool component.
@@ -88,15 +46,6 @@ export interface WorkerPoolOptions {
     nodeSubnetIds?: pulumi.Input<pulumi.Input<string>[]>;
 
     /**
-     * Optional mappings from AWS IAM roles to Kubernetes users and groups.
-     */
-    roleMappings?: pulumi.Input<pulumi.Input<RoleMapping>[]>;
-
-     /**
-     * Optional mappings from AWS IAM users to Kubernetes users and groups.
-     */
-    userMappings?: pulumi.Input<pulumi.Input<UserMapping>[]>;
-    /**
      * The security group associated with the EKS cluster.
      */
     clusterSecurityGroup: aws.ec2.SecurityGroup;
@@ -112,9 +61,14 @@ export interface WorkerPoolOptions {
     instanceType?: pulumi.Input<aws.ec2.InstanceType>;
 
     /**
-     * The instance role to use for all nodes in this workder pool.
+     * The instance profile to use for all nodes in this workder pool.
      */
-    instanceRole?: pulumi.Input<aws.iam.Role>;
+    instanceProfile: pulumi.Input<aws.iam.InstanceProfile>;
+
+    /**
+     * The eks node access config map used by worker nodes
+     */
+    eksNodeAccess: k8s.core.v1.ConfigMap;
 
     /**
      * The security group to use for all nodes in this workder pool.
@@ -173,11 +127,6 @@ export interface WorkerPoolOptions {
  */
 export class WorkerPool extends pulumi.ComponentResource {
     /**
-     * The service role used by the EKS cluster.
-     */
-    public readonly instanceRole: pulumi.Output<aws.iam.Role>;
-
-    /**
      * The security group for the cluster's nodes.
      */
     public readonly nodeSecurityGroup: aws.ec2.SecurityGroup;
@@ -198,35 +147,17 @@ export class WorkerPool extends pulumi.ComponentResource {
             throw new pulumi.RunError("a 'kubernetes' provider must be specified for a 'WorkerPool'");
         }
         const pool = createWorkerPool(name, args, this, k8sProvider);
-        this.instanceRole = pool.instanceRole;
         this.nodeSecurityGroup = pool.nodeSecurityGroup;
         this.registerOutputs(undefined);
     }
 }
 
 export interface WorkerPoolData {
-    instanceRole: pulumi.Output<aws.iam.Role>;
     nodeSecurityGroup: aws.ec2.SecurityGroup;
     cfnStack: aws.cloudformation.Stack;
 }
 
 export function createWorkerPool(name: string, args: WorkerPoolOptions, parent: pulumi.ComponentResource,  k8sProvider: k8s.Provider): WorkerPoolData {
-    let instanceRole: pulumi.Output<aws.iam.Role>;
-    if (args.instanceRole) {
-        instanceRole = pulumi.output(args.instanceRole);
-    } else {
-        instanceRole = (new ServiceRole(`${name}-instanceRole`, {
-            service: "ec2.amazonaws.com",
-            managedPolicyArns: [
-                "arn:aws:iam::aws:policy/AmazonEKSWorkerNodePolicy",
-                "arn:aws:iam::aws:policy/AmazonEKS_CNI_Policy",
-                "arn:aws:iam::aws:policy/AmazonEC2ContainerRegistryReadOnly",
-            ],
-        }, { parent: parent })).role;
-    }
-    // Create the instance role we'll use for worker nodes.
-    const instanceRoleARN = instanceRole.apply(r => r.arn);
-
     let nodeSecurityGroup: aws.ec2.SecurityGroup;
     if (args.nodeSecurityGroup) {
         nodeSecurityGroup = args.nodeSecurityGroup;
@@ -268,46 +199,6 @@ export function createWorkerPool(name: string, args: WorkerPoolOptions, parent: 
             }),
         }, { parent: parent });
     }
-
-    // Enable access to the EKS cluster for worker nodes.
-    const instanceRoleMapping: RoleMapping = {
-        roleArn: instanceRoleARN,
-        username: "system:node:{{EC2PrivateDNSName}}",
-        groups: ["system:bootstrappers", "system:nodes"],
-    };
-    const roleMappings = pulumi.all([pulumi.output(args.roleMappings || []), instanceRoleMapping])
-        .apply(([mappings, instanceMapping]) => {
-            return jsyaml.safeDump([...mappings, instanceMapping].map(m => ({
-                rolearn: m.roleArn,
-                username: m.username,
-                groups: m.groups,
-            })));
-        });
-    const nodeAccessData: any = {
-        mapRoles: roleMappings,
-    };
-    if (args.userMappings !== undefined) {
-        nodeAccessData.mapUsers = pulumi.output(args.userMappings).apply(mappings => {
-            return jsyaml.safeDump(mappings.map(m => ({
-                userarn: m.userArn,
-                username: m.username,
-                groups: m.groups,
-            })));
-        });
-    }
-    const eksNodeAccess = new k8s.core.v1.ConfigMap(`${name}-nodeAccess`, {
-        apiVersion: "v1",
-        metadata: {
-            name: `aws-auth-${name}`,
-            namespace: "kube-system",
-        },
-        data: nodeAccessData,
-    }, { parent: parent, provider: k8sProvider });
-
-    // Create the cluster's worker nodes.
-    const instanceProfile = new aws.iam.InstanceProfile(`${name}-instanceProfile`, {
-        role: instanceRole,
-    }, { parent: parent });
 
     const eksClusterIngressRule = new aws.ec2.SecurityGroupRule(`${name}-eksClusterIngressRule`, {
         description: "Allow pods to communicate with the cluster API Server",
@@ -370,7 +261,7 @@ ${customUserData}
         associatePublicIpAddress: true,
         imageId: amiId,
         instanceType: args.instanceType || "t2.medium",
-        iamInstanceProfile: instanceProfile.id,
+        iamInstanceProfile: args.instanceProfile,
         keyName: keyName,
         securityGroups: [ nodeSecurityGroupId ],
         rootBlockDevice: {
@@ -416,10 +307,9 @@ ${customUserData}
     const cfnStack = new aws.cloudformation.Stack(`${name}-nodes`, {
         name: cfnStackName,
         templateBody: cfnTemplateBody,
-    }, { parent: parent, dependsOn: [eksNodeAccess, args.vpcCni] });
+    }, { parent: parent, dependsOn: [args.eksNodeAccess, args.vpcCni] });
 
     return {
-        instanceRole: instanceRole,
         nodeSecurityGroup: nodeSecurityGroup,
         cfnStack: cfnStack,
     };
